inint
c

+++ galileo-server/galileo-i18n-server/src/main/java/org/armada/galileo/i18n_server/dal/bo/I18nDictionaryValueBO.java
@@ -18,5 +18,5 @@ import java.util.Set;
 
 public interface I18nDictionaryValueBO extends BaseBO<I18nDictionaryValue, I18nDictionaryValueDTO> {
 
-    List<Integer> getUnfinishedDictionaryByAppIdAndLocals(Set<String> keys, Integer appId, Integer size);
+    List<Long> getUnfinishedDictionaryByAppIdAndLocals(Set<String> keys, Long appId, Long size);
 }
