+++ galileo-server/galileo-i18n-server/src/main/java/org/armada/galileo/i18n_server/dal/vo/DictQueryVO.java
@@ -11,7 +11,7 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 public class DictQueryVO {
 
-    Integer appId;
+    Long appId;
 
     private String appCode;
